shadow-dom/shadow-trees/nested-shadow-trees/cannot-access-from-external-dom.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Shadow DOM</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
    <h1>nested shadow trees level 2</h1>
    <div id="level0"></div>
    <div id="log"></div>
    <template id="level1-template">
      <div id="level1">
        <div id="level2-root"></div>
      </div>

    </template>
    <template id="level2-template">
      <div id="level2">
      </div>
    </template>
    <script>
        test(function() {
            var level0 = document.getElementById('level0'),
                shadowLevel1 = level0.webkitCreateShadowRoot(),
                shadowLevel2,
                level1 = document.getElementById('level1-template').content,
                level2 = document.getElementById('level2-template').content;

            shadowLevel1.appendChild(level1);

            shadowLevel2 = shadowLevel1.getElementById('level2-root').webkitCreateShadowRoot();

            shadowLevel2.appendChild(level2);

            assert_equals(level0.querySelector('#level1'), null);
            assert_equals(level0.querySelector('#level2'), null);
            assert_equals(shadowLevel1.querySelector('#level2'), null);
            assert_not_equals(shadowLevel1.querySelector('#level1'), null);
            assert_not_equals(shadowLevel2.querySelector('#level2'), null);
        }, 'nested shadow tree cannot access from external');
    </script>
 </body>
</html>
